+++ GeSIR_app/api/viewsProvider.py
@@ -1,0 +1,38 @@
+from rest_framework import generics
+from .models import Proveedor
+from .serializers import ProveedorSerializer, CoordenadasSerializer
+from rest_framework.response import Response
+from rest_framework.permissions import IsAuthenticated
+
+
+class ProveedorListCreateView(generics.ListCreateAPIView):
+    permission_classes = [IsAuthenticated]
+    queryset = Proveedor.objects.all()
+    serializer_class = ProveedorSerializer
+ 
+class ProveedorUpdateView(generics.RetrieveUpdateAPIView):
+    permission_classes = [IsAuthenticated]
+    queryset = Proveedor.objects.all()
+    serializer_class = ProveedorSerializer
+    partial = True
+    
+class ProveedorDeleteView(generics.DestroyAPIView):
+    permission_classes = [IsAuthenticated]
+    queryset = Proveedor.objects.all()
+    serializer_class = ProveedorSerializer
+    
+    def destroy(self, request, *args, **kwargs):
+        instance = self.get_object()
+        instance.delete()
+        return Response(print("Delete Proveedor"))
+
+class ProveedorDetailView(generics.RetrieveUpdateDestroyAPIView):
+    permission_classes = [IsAuthenticated]
+    queryset = Proveedor.objects.all()
+    serializer_class = ProveedorSerializer
+
+
+class AllProveedoresListView(generics.ListAPIView):
+    permission_classes = [IsAuthenticated]
+    queryset = Proveedor.objects.all()
+    serializer_class = ProveedorSerializer
